+++ src/com/github/dreamrec/FrameDecoderListener.java
@@ -7,7 +7,7 @@ public interface FrameDecoderListener {
     void setFrameCounter250(int value);
     void setCh1Value(int value);
     void setCh2Value(int value);
-    void setAcc1Value(int value);
-    void setAcc2Value(int value);
-    void setAcc3Value(int value);
+    void setAcc1Value(short value);
+    void setAcc2Value(short value);
+    void setAcc3Value(short value);
 }
