+++ src/com/github/dreamrec/LoPassFilter.java
@@ -3,25 +3,25 @@ package com.github.dreamrec;
 /**
  *
  */
-public class LoPassFilter extends AbstractFilter<Integer>{
+public class LoPassFilter extends AbstractFilter<Short>{
 
     private int bufferSize;
 
-    public LoPassFilter(int bufferSize, Filter<Integer> inputData) {
+    public LoPassFilter(int bufferSize, Filter<Short> inputData) {
         super(inputData);
         this.bufferSize = bufferSize;
     }
 
     @Override
-    protected Integer doFilter(int index) {
-        long sum = 0;
+    protected Short doFilter(int index) {
+        int sum = 0;
         if(index<bufferSize-1){
             return inputData.get(index);
         }else {
             for (int i = 0; i < bufferSize; i++) {
                  sum+=inputData.get(index - i);
             }
-            return (int)(sum/bufferSize);
+            return (short)(sum/bufferSize);
         }
     }
 }
